+++ pkg/local_object_storage/shard/select.go
@@ -3,6 +3,8 @@ package shard
 import (
 	"fmt"
 
+	meta "github.com/nspcc-dev/neofs-node/pkg/local_object_storage/metabase"
+	"github.com/nspcc-dev/neofs-sdk-go/client"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
 	"github.com/nspcc-dev/neofs-sdk-go/object"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
@@ -28,4 +30,18 @@ func (s *Shard) Select(cnr cid.ID, filters object.SearchFilters) ([]oid.Address,
 	}
 
 	return addrs, nil
+}
+
+// Search performs Search op on the underlying metabase if it is not disabled.
+func (s *Shard) Search(cnr cid.ID, fs object.SearchFilters, attrs []string, cursor *meta.SearchCursor, count uint16) ([]client.SearchResultItem, *meta.SearchCursor, error) {
+	s.m.RLock()
+	defer s.m.RUnlock()
+	if s.info.Mode.NoMetabase() {
+		return nil, nil, ErrDegradedMode
+	}
+	res, cursor, err := s.metaBase.Search(cnr, fs, attrs, cursor, count)
+	if err != nil {
+		return nil, nil, fmt.Errorf("call metabase: %w", err)
+	}
+	return res, cursor, nil
 }